led stimulator initial version

+++ engine/vision_experiment/stimulation_library.py
@@ -309,8 +309,8 @@ class Stimulations(experiment_control.StimulationControlHelper):#, screen.Screen
             fns = os.listdir(path)
             fns.sort()
             if length>0:
-                length_f=(self.config.SCREEN_EXPECTED_FRAME_RATE*length)
-                offset_f=(self.config.SCREEN_EXPECTED_FRAME_RATE*offset)
+                length_f=int(self.config.SCREEN_EXPECTED_FRAME_RATE*length)
+                offset_f=int(self.config.SCREEN_EXPECTED_FRAME_RATE*offset)
                 fns=fns[offset_f:offset_f+length_f]
             if len([f for f in fns if os.path.splitext(f)[1] not in ['.png', '.bmp', '.jpg']])>0:
                  raise RuntimeError('{0} folder contains non image files, please remove them!'.format(path))
